fix: 일일 재고 리셋 consumer 위치 수정
